@@ -1,6 +1,6 @@
 module ocm.software/ocm
 
-go 1.23.6
+go 1.24.0
 
 replace github.com/spf13/cobra => github.com/open-component-model/cobra v0.0.0-20230329075350-b1fd876abfb9
 
