@@ -1,3 +1,3 @@
 module ocm.software/ocm/hack/brew
 
-go 1.23.2
+go 1.24.0
