@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"golang.org/x/exp/slices"
+	"slices"
 
 	"ocm.software/ocm/api/datacontext/action/api"
 )
