@@ -1,8 +1,9 @@
 package vault
 
 import (
+	"slices"
+
 	"github.com/mandelsoft/goutils/optionutils"
-	"golang.org/x/exp/slices"
 
 	"ocm.software/ocm/api/utils"
 )
