@@ -1,11 +1,11 @@
 package pubsub
 
 import (
+	"maps"
 	"sync"
 
 	"github.com/mandelsoft/goutils/errors"
 	"github.com/mandelsoft/goutils/general"
-	"golang.org/x/exp/maps"
 
 	"ocm.software/ocm/api/ocm/cpi"
 )
