@@ -3,12 +3,12 @@ package internal
 import (
 	"encoding/json"
 	"fmt"
+	"slices"
 	"sync"
 
 	"github.com/mandelsoft/goutils/errors"
 	"github.com/mandelsoft/goutils/maputils"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/exp/slices"
 
 	"ocm.software/ocm/api/utils/runtime"
 )
