@@ -9,7 +9,6 @@ import (
 	"github.com/mandelsoft/goutils/errors"
 	"github.com/mandelsoft/goutils/sliceutils"
 	"github.com/mandelsoft/vfs/pkg/vfs"
-	"golang.org/x/exp/maps"
 
 	"ocm.software/ocm/api/ocm/compdesc"
 	metav1 "ocm.software/ocm/api/ocm/compdesc/meta/v1"
@@ -139,8 +138,11 @@ func (v *verifiedStore) Entries() []common.NameVersion {
 	v.lock.Lock()
 	defer v.lock.Unlock()
 
-	return sliceutils.Transform(maps.Values(v.storage.ComponentVersions),
-		func(cd *StorageEntry) common.NameVersion { return common.VersionedElementKey(cd.Descriptor) })
+	entries := make([]common.NameVersion, 0, len(v.storage.ComponentVersions))
+	for _, entry := range v.storage.ComponentVersions {
+		entries = append(entries, common.VersionedElementKey(entry.Descriptor))
+	}
+	return entries
 }
 
 func (v *verifiedStore) Add(cd *compdesc.ComponentDescriptor, signatures ...string) {
